refactor: move dashboard and metric-sidebar
Move components/dashboard and components/metric-sidebar to nerdlets/mande-nerdlet/components
