+++ nerdlets/mande-nerdlet/components/metric-sidebar/MetricSidebar.js
