+++ nerdlets/mande-nerdlet/components/metric-sidebar/style.scss
