+++ nerdlets/mande-nerdlet/styles.scss
@@ -4,8 +4,8 @@
 @import '../shared/scss/empty-view.scss';
 @import "../shared/components/metric/styles.scss";
 @import "./components/category-menu/style.scss";
-@import "../../components/metric-sidebar/style.scss";
-@import "../../components/dashboard/styles";
+@import "./components/metric-sidebar/style.scss";
+@import "./components/dashboard/styles";
 @import "./components/category-detail/styles";
 
 // ==============================================================
